backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
pymupdf